+++ scripts/run_tests_within_container.sh
@@ -5,7 +5,7 @@ export KB_AUTH_TOKEN=`cat /kb/module/work/token`
 export PYTHONPATH=$script_dir/../lib:$PATH:$PYTHONPATH
 
 # Set TEST_PATH to run a specific test. Eg: TEST_PATH=test.core.update_taxon_assignments_test
-export TEST_PATH=.
+export TEST_PATH=test.problematic_tests.impl_test
 
 cd $script_dir/../test
 python -m nose --with-coverage --cover-package=GenomeFileUtil --cover-html --cover-html-dir=/kb/module/work/test_coverage --cover-xml --cover-xml-file=/kb/module/work/test_coverage/coverage.xml --nocapture --nologcapture $TEST_PATH
