+++ test/problematic_tests/save_genome_test.py
@@ -5,7 +5,6 @@ import json  # noqa: F401
 import os  # noqa: F401
 import shutil
 import time
-import traceback
 import unittest
 import urllib.error
 import urllib.parse
@@ -14,7 +13,6 @@ from configparser import ConfigParser
 from os import environ
 
 import requests  # noqa: F401
-from pytest import raises
 
 from installed_clients.AssemblyUtilClient import AssemblyUtil
 from installed_clients.DataFileUtilClient import DataFileUtil
@@ -24,19 +22,6 @@ from GenomeFileUtil.GenomeFileUtilServer import MethodContext
 from GenomeFileUtil.authclient import KBaseAuth as _KBaseAuth
 from GenomeFileUtil.core.GenomeInterface import GenomeInterface
 from installed_clients.WorkspaceClient import Workspace as workspaceService
-
-
-def assert_exception_correct(got: Exception, expected: Exception):
-    """
-    Compare raised exception with expected exception.
-
-    Args:
-        got (Exception): Exception received
-        expected (Exception): Exception expected
-    """
-    err = "".join(traceback.TracebackException.from_exception(got).format())
-    assert got.args == expected.args, err
-    assert type(got) == type(expected)
 
 
 class SaveGenomeTest(unittest.TestCase):
@@ -153,9 +138,6 @@ class SaveGenomeTest(unittest.TestCase):
     def getContext(self):
         return self.__class__.ctx
 
-    def getCfg(self):
-        return self.__class__.cfg
-
     def start_test(self):
         testname = inspect.stack()[1][3]
         print(('\n*** starting test: ' + testname + ' **'))
@@ -209,34 +191,6 @@ class SaveGenomeTest(unittest.TestCase):
                   'hidden': True}
         ret = self.getImpl().save_one_genome(self.ctx, params)[0]
         self.check_save_one_genome_output(ret, genome_name)
-
-    def test_invalid_catalog_param_type(self):
-        self.start_test()
-
-        MAX_THREADS = "MAX_THREADS"
-        THREADS_PER_CPU = "THREADS_PER_CPU"
-
-        # max_threads type check fails
-        os.environ["KBASE_SECURE_CONFIG_PARAM_MAX_THREADS"] = "10.5"
-        os.environ["KBASE_SECURE_CONFIG_PARAM_THREADS_PER_CPU"] = "2.5"
-
-        with raises(Exception) as got:
-            GenomeFileUtil(self.getCfg())
-        assert_exception_correct(
-            got.value,
-            ValueError(f"{MAX_THREADS} must be of type {int.__name__}")
-        )
-
-        # threads_per_cpu type check fails
-        os.environ["KBASE_SECURE_CONFIG_PARAM_MAX_THREADS"] = "10"
-        os.environ["KBASE_SECURE_CONFIG_PARAM_THREADS_PER_CPU"] = "2.8e"
-
-        with raises(Exception) as got:
-            GenomeFileUtil(self.getCfg())
-        assert_exception_correct(
-            got.value,
-            ValueError(f"{THREADS_PER_CPU} must be of type {float.__name__}")
-        )
 
     def test_GenomeInterface_check_dna_sequence_in_features(self):
         # no feature in genome
